// src/types/api.ts
export type GuildMember = {
    user: User;
	nick?: string;
	avatar?: string;
	banner?: string;
	roles: string[];
	joined_at: string;
	premium_since?: string;
	deaf: boolean;
	mute: boolean;
	pending?: boolean;
	permissions?: string;
	communication_disabled_until?: string;
	avatar_decoration_data?: UserAvatarDecorationData | null;
}
// ...
export type GuildRole = {
    id: string;
    name: string;
    color: number;
    hex_color?: string;
    hoist: boolean;
    icon_url?: string;
    unicode_emoji?: string;
    position: number;
    permissions: number;
    managed: boolean;
    mentionable: boolean;
    tags?: GuildRoleTags;
    flags: GuildRoleFlags[];
}
// ...
export type Channel = {
    id: string;
    type: ChannelTypes;
    guild_id: string;
    position: number;
    permission_overwrites: ChannelOverwrite[];
    name: string;
    topic?: string;
    nsfw: boolean;
    last_message_id?: string;
    /**
     * @requires type = 2 | 13
     */
    bitrate?: number;
    /**
     * @requires type = 2 | 13
     */
    user_limit?: number;
    rate_limit_per_user?: number;
    /**
     * @requires type = 10 | 11 | 12
     */
    owner_id?: string;
    parent_id?: string;
    last_pin_timestamp?: string;
    /**
     * @requires type = 2 | 13
     */
    rtc_region?: string;
    /**
     * @requires type = 2 | 13
     */
    video_quality_mode?: number;
    /**
     * @requires type = 10 | 11 | 12
     */
    message_count?: number;
    /**
     * @requires type = 10 | 11 | 12
     */
    member_count?: number;
    /**
     * @requires type = 10 | 11 | 12
    */
    thread_metadata?: {
       archived: boolean;
       auto_archive_duration: number;
       archive_timestamp: string;
       locked: boolean;
       invitable?: boolean;
    };
    default_auto_archive_duration?: number;
    permissions?: string;
    flags?: number;
    /**
     * @requires type = 10 | 11 | 12
     */
    total_message_sent?: number;
    /**
     * @requires type = 15 | 19
     */
    available_tags?: ChannelTag[];
    /**
     * @requires type = 15 | 19
     */
    applied_tags?: string[];
    /**
     * @requires type = 15 | 19
     */
    default_reaction_emoji?: ChannelDefaultReaction;
    default_thread_rate_limit_per_user?: number;
    /**
     * @requires type = 15 |19
     */
    default_sort_order?: number;
    /**
     * @requires type = 15
     */
    default_forum_layout?: number;
}
export type ChannelOverwrite = {
    id: string;
    type: ChannelOverwriteTypes;
    allow: string | number;
    deny: string | number;
}
export enum ChannelOverwriteTypes {
    "ROLE" = 0,
    "MEMBER" = 1,
}
export enum ChannelTypes {
    "GUILD_TEXT" = 0,
    "DM" = 1,
    "GUILD_VOICE" = 2,
    "GROUP_DM" = 3,
    "GUILD_CATEGORY" = 4,
    "GUILD_ANNOUNCEMENT" = 5,
    "ANNOUNCEMENT_THREAD" = 10,
    "PUBLIC_THREAD" = 11,
    "PRIVATE_THREAD" = 12,
    "GUILD_STAGE_VOICE" = 13,
    "GUILD_DIRECTORY" = 14,
    "GUILD_FORUM" = 15,
    "GUILD_MEDIA" = 19,

}
type ChannelTag = {
    id: string;
    name: string;
    moderated: boolean;
    emoji_id: string | null;
    emoji_name: string | null;
}
type ChannelDefaultReaction = {
    emoji_id: string | null;
    emoji_name: string | null;
}

export type Message = {
    id: string;
    channel_id: string;
    author: User;
    content: string;
    timestamp: string;
    edited_timestamp: string | null;
    tts: boolean;
    mention_everyone: boolean;
    mentions: User[];
    mention_roles: string[];
    mention_channels: MessageChannelMention[];
    attachments: MessageAttachment[];
    embeds: MessageEmbed[];
    reactions: MessageReaction[];
    nonce?: string | null;
    pinned: boolean;
    webhook_id?: string;
    type: number;
    activity?: MessageActivity;
    application_id?: string;
    message_reference?: MessageReference;
    flags?: MessageFlags;
    referenced_message?: Message | null;
    interaction?: MessageInteractionMetadata;
    thread?: Channel;
    components?: MessageComponent[];
    sticker_items?: StickerItem[];
    stickers?: Sticker[];
    position?: number;
    role_subscription_data?: MessageRoleSubscriptionData;
    resolved?: MessageResolved;
    poll?: Poll;
    call?: MessageCall;
}
// ...
export enum MessageComponentTypes {
    "ACTION_ROW" = 1,
    "BUTTON" = 2,
    "STRING_SELECT" = 3,
    "USER_SELECT" = 5,
    "ROLE_SELECT" = 6,
    "MENTIONABLE_SELECT" = 7,
    "CHANNEL_SELECT" = 8,
    "SECTION" = 9,
    "TEXT_DISPLAY" = 10,
    "THUMBNAIL" = 11,
    "MEDIA_GALLERY" = 12,
    "FILE" = 13,
    "SEPARATOR" = 14,
    "CONTAINER" = 17,
}
type MessageComponent = {
    type: MessageComponentTypes.ACTION_ROW;
    components: MessageComponentActionRow[];
}
type MessageComponentActionRow = MessageComponentButton | MessageComponentSelect;
// ...
type MessageActivity = {
    type: MessageActivityTypes;
    party_id?: string;
}
export enum MessageActivityTypes {
    "JOIN" = 1,
    "SPECTATE" = 2,
    "LISTEN" = 3,
    "JOIN_REQUEST" = 5,
}
export enum MessageFlags {
    "CROSSPOSTED" = 1 << 0,
    "IS_CROSSPOST" = 1 << 1,
    "SUPPRESS_EMBEDS" = 1 << 2,
    "SOURCE_MESSAGE_DELETED" = 1 << 3,
    "URGENT" = 1 << 4,
    "HAS_THREAD" = 1 << 5,
    "EPHEMERAL" = 1 << 6,
    "LOADING" = 1 << 7,
    "FAILED_TO_MENTION_SOME_ROLES_IN_THREAD" = 1 << 8,
    "SUPRESS_NOTIFICATIONS" = 1 << 12,
    "IS_VOICE_MESSAGE" = 1 << 13,
    "HAS_SNAPSHOT" = 1 << 14,
}
type MessageInteractionMetadata = (MessageComponentInteractionMetadata | MessageCommandInteractionMetadata | MessageModalSubmitInteractionMetadata) & {
    id: string;
    user: User;
    authorizing_integration_owners: InteractionIntegrationOwners;
    original_response_mnessage_id?: string;
} 
type MessageComponentInteractionMetadata = {
    type: InteractionTypes.MESSAGE_COMPONENT;
    interacted_message_id: string;
}
type MessageCommandInteractionMetadata = {
    type: InteractionTypes.APPLICATION_COMMAND;
    target_message_id?: string;
    target_user?: User;
}
type MessageModalSubmitInteractionMetadata = {
    type: InteractionTypes.MODAL_SUBMIT;
    triggering_interaction_metadata: MessageComponentInteractionMetadata | MessageCommandInteractionMetadata;
}
type MessageCall = {
    participants: string[];
    ended_timestamp?: string | null;
}
type MessageReference = {
    type: MessageReferenceTypes;
    message_id?: string;
    channel_id?: string;
    guild_id?: string;
    fail_if_not_exists?: boolean;
}
export enum MessageReferenceTypes {
    "DEFAULT" = 0,
    "FORWARD" = 1,
}
type MessageReaction = {
    count: number;
    count_details: {
        burst: number;
        normal: number;
    };
    me: boolean;
    me_burst: boolean;
    emoji: Emoji;
    burst_colors: string[];
}
type MessageEmbed = MessageEmbedPoll & {
    title?: string;
    type?: MessageEmbedTypes;
    description?: string;
    url?: string;
    timestamp?: string;
    color?: number;
    footer?: MessageEmbedFooter;
    image?: MessageEmbedImage;
    thumbnail?: MessageEmbedThumbnail;
    video?: MessageEmbedVideo;
    provider?: MessageEmbedProvider;
    author?: MessageEmbedAuthor;
    fields?: MessageEmbedField[];
}
export enum MessageEmbedTypes {
    "RICH" = "rich",
    "IMAGE" = "image",
    "VIDEO" = "video",
    "GIFV" = "gifv",
    "ARTICLE" = "article",
    "LINK" = "link",
    "POLL_RESULT" = "poll_result",
}
type MessageEmbedFooter = {
    text: string;
    icon_url?: string;
}
type MessageEmbedImage = {
    url: string;
}
type MessageEmbedThumbnail = {
    url: string;
}
type MessageEmbedVideo = {
    url: string;
}
type MessageEmbedProvider = {
    name: string;
    url?: string;
}
type MessageEmbedAuthor = {
    name: string;
    url?: string;
    icon_url?: string;
}
type MessageEmbedField = {
    name: string;
    value: string;
    inline?: boolean;
}
type MessageEmbedPoll = {
    type: MessageEmbedTypes.POLL_RESULT;
    poll_question_text: string;
    victor_answer_votes: number;
    total_votes: number;
    victor_answer_id?: string;
    victor_answer_text?: string;
    victor_answer_emoji_id?: string;
    victor_answer_emoji_name?: string;
    victor_answer_emoji_animated?: boolean;
}
export type MessageAttachment = {
    id: string;
    filename: string;
    title?: string;
    description?: string;
    content_type?: string;
    size: number;
    url: string;
    proxy_url: string;
    height?: number;
    width?: number;
    ephemeral?: boolean;
    duration_secs?: number;
    waveform?: string;
    flags?: MessageAttachmentFlags;
}
export enum MessageAttachmentFlags {
    "IS_REMIX" = 1 << 2.
}
type MessageChannelMention = {
    id: string;
    guild_id: string;
    type: ChannelTypes;
    name: string;
}
type MessageAllowedMentions = {
    parse?: string[];
    roles?: string[];
    users?: string[];
    replied_user?: boolean;
}
type MessageRoleSubscriptionData = {
    role_subscription_listing_id: string;
    tier_name: string;
    total_months_subscribed: number;
    is_renewal: boolean;
}
type MessageResolved = {
    users?: Map<string, User>;
    members?: Map<string, GuildMember>;
    roles?: Map<string, GuildRole>;
    channels?: Map<string, Channel>;
    messages?: Map<string, Message>;
    attachments?: Map<string, MessageAttachment>;
}

export type Emoji = {
    id: string | null;
    name: string | null;
    roles?: string[];
    user?: User;
    require_colons?: boolean;
    managed?: boolean;
    animated?: boolean;
    available?: boolean;
}

export type Poll = {
    question: PollMedia;
    answers: PollAnswer[];
    expiry: string | null;
    allow_multiselect: boolean;
    layout_type: PollLayoutType;
    results?: PollResults;
}
export enum PollLayoutType {
    "DEFAULT" = 1,
}
type PollAnswer = {
    answer_id: string;
    poll_media: PollMedia;
}
type PollMedia = {
    text: string;
    emoji?: Emoji;
}
type PollResults = {
    is_finalized: boolean;
    answer_counts: PollAnswerCount[];
}
type PollAnswerCount = {
    id: string;
    count: number;
    me_voted: boolean;
}

export type Sticker = {
	id: string;
	pack_id?: string;
	name: string;
	description: string | null;
	tags: string;
	asset: string;
	type: number;
	format_type: number;
	available: boolean;
	guild_id: string | null;
	user: User | null;
	sort_value: number;
}
export type StickerItem = {
    id: string;
    name: string;
    format_type: StickerFormatType;
}
export enum StickerFormatType {
	"PNG" = 1,
	"APNG" = 2,
	"LOTTIE" = 3,
    "GIF" = 4
}
export enum StickerType {
	"STANDARD" = 1,
	"GUILD" = 2,
}

export type User = {
	id: string;
	username: string;
	discriminator: string;
	global_name: string | null;
	avatar: string | null;
	bot?: boolean;
	system?: boolean;
	mfa_enabled?: boolean;
	banner?: string | null;
	accent_color?: number | null;
	locale?: string;
	verified?: boolean;
	email?: string | null;
	flags?: number;
	premium_type?: UserPremiumType;
	public_flags?: number;
	avatar_decoration_data?: UserAvatarDecorationData | null;
    primary_guild?: UserPrimaryGuild | null;
}
type UserAvatarDecorationData = {
	asset: string;
	sku_id: string;
}
type UserPrimaryGuild = {
    identity_guild_id: string;
    identity_enabled: boolean;
    tag: string;
    badge: string;
}
// ...
export enum UserPremiumType {
	"NONE" = 0,
	"NITRO_CLASSIC" = 1,
	"NITRO" = 2,
	"NITRO_BASIC" = 3,
}